practice/3_document/pyproject.toml
[project]
name = "3-document"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "docx2pdf>=0.1.8",
    "docxtpl>=0.20.2",
    "fastapi>=0.121.2",
    "pydantic-settings>=2.12.0",
    "pydentic>=0.0.1.dev3",
    "python-docx>=1.2.0",
    "uvicorn>=0.38.0",
]
